@@ -7,13 +7,7 @@ use Illuminate\Http\Request;
 
 class DetachAttachment
 {
-    /**
-     * Delete an attachment from the field.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return void
-     */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request): void
     {
         Attachment::where('url', $request->src)
                         ->get()
