@@ -7,13 +7,7 @@ use Illuminate\Http\Request;
 
 class DiscardPendingAttachments
 {
-    /**
-     * Discard pendings attachments on the field.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return void
-     */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request): void
     {
         PendingAttachment::where('draft_id', $request->draftId)
                     ->get()
