@@ -2,18 +2,13 @@
 
 namespace Froala\NovaFroalaField\Http\Controllers;
 
+use Illuminate\Http\JsonResponse;
 use Laravel\Nova\Http\Controllers\TrixAttachmentController;
 use Laravel\Nova\Http\Requests\NovaRequest;
 
 class FroalaToTrixAttachmentAdapterController extends TrixAttachmentController
 {
-    /**
-     * Store an attachment for a Trix field.
-     *
-     * @param  \Laravel\Nova\Http\Requests\NovaRequest  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(NovaRequest $request)
+    public function store(NovaRequest $request): JsonResponse
     {
         $response = parent::store($request);
 
