@@ -6,17 +6,13 @@ use Froala\NovaFroalaField\Models\PendingAttachment;
 
 class PruneStaleAttachments
 {
-    /**
-     * Prune the stale attachments from the system.
-     *
-     * @return void
-     */
-    public function __invoke()
+    public function __invoke(): void
     {
-        PendingAttachment::where('created_at', '<=', now()->subDays(1))
-                    ->orderBy('id', 'desc')
-                    ->chunk(100, function ($attachments) {
-                        $attachments->each->purge();
-                    });
+        PendingAttachment::query()
+            ->where('created_at', '<=', now()->subDays(1))
+            ->orderBy('id', 'desc')
+            ->chunk(100, function ($attachments) {
+                $attachments->each->purge();
+            });
     }
 }
